requirements.txt
Flask==2.3.2
psycopg2-binary==2.9.9  # Explicitly specify version
python-dotenv==1.1.1
Flask-Cors==4.0.0
shapely==2.0.2
geojson==3.1.0 
gunicorn==21.2.0